update folders + flags

+++ client_datalink/src/routes/machine/[id]/_components/FoldersModal.svelte
@@ -2,7 +2,15 @@
 	import Button from '$lib/components/ui/button/button.svelte';
 	import * as Dialog from '$lib/components/ui/dialog/index';
 	import InputGroup from '$lib/components/atoms/InputGroup.svelte';
-	import { BookmarkPlusIcon, FolderPen, FoldersIcon, PinIcon, PinOff, Trash } from '@lucide/svelte';
+	import {
+		BookmarkPlusIcon,
+		FolderPen,
+		FoldersIcon,
+		PinIcon,
+		PinOff,
+		Trash,
+		UserStar
+	} from '@lucide/svelte';
 	import {
 		useCreateNewUserFolder,
 		useDeleteFolder,
@@ -12,8 +20,13 @@
 	import { page } from '$app/state';
 	import LoadingTemplate from '$lib/components/atoms/LoadingTemplate.svelte';
 	import NodataTemplate from '$lib/components/atoms/NodataTemplate.svelte';
+	import { userStore } from '$lib/stores/UserStore.svelte';
 
-	let { tagName, isText }: { tagName?: string; isText?: boolean } = $props();
+	let {
+		tagName,
+		isOtherFolder = false,
+		isText
+	}: { tagName?: string; isOtherFolder: boolean; isText?: boolean } = $props();
 
 	const machineId = $derived(page.params.id!);
 	const createNameMutate = useCreateNewUserFolder();
@@ -22,7 +35,9 @@
 	const allFoldersBm = useGetAllUserFolders(() => machineId);
 
 	let open = $state(false);
-	let tab = $state<'existbm' | 'createnewbm'>('existbm');
+	let tab = $derived<'existbm' | 'createnewbm' | 'otherfolders'>(
+		!isOtherFolder ? 'existbm' : 'otherfolders'
+	);
 	let newFolderName = $state('');
 
 	let sortedFolders = $derived(
@@ -93,7 +108,11 @@
 		</Dialog.Header>
 
 		<div class="mb-12">
-			<section class="grid grid-cols-2 gap-2 p-1 bg-secondary rounded-lg my-4">
+			<section
+				class="grid {userStore.isAdmin
+					? 'grid-cols-3'
+					: 'grid-cols-2'} gap-3 bg-secondary rounded-lg my-4"
+			>
 				<Button
 					variant={tab === 'existbm' ? 'default' : 'ghost'}
 					size="sm"
@@ -102,13 +121,23 @@
 				>
 					<FoldersIcon /><span>Your folders</span>
 				</Button>
+				{#if userStore.isAdmin}
+					<Button
+						variant={tab === 'otherfolders' ? 'default' : 'ghost'}
+						size="sm"
+						onclick={() => (tab = 'otherfolders')}
+						class="rounded-xl!"
+					>
+						<UserStar /><span>Other folders</span>
+					</Button>
+				{/if}
 				<Button
 					variant={tab === 'createnewbm' ? 'default' : 'ghost'}
 					size="sm"
 					onclick={() => (tab = 'createnewbm')}
 					class="rounded-xl!"
 				>
-					<BookmarkPlusIcon /><span>Create new folder</span>
+					<BookmarkPlusIcon /><span>Create new</span>
 				</Button>
 			</section>
 
@@ -122,38 +151,40 @@
 						{:else if allFoldersBm.isError}
 							<p>{allFoldersBm.error}</p>
 						{:else if allFoldersBm.isSuccess}
-							{#if allFoldersBm.data?.length === 0}
+							{#if allFoldersBm.data.filter((f) => f.userId === userStore.user?.id).length === 0}
 								<NodataTemplate text="No folders..." size="sm" />
 							{/if}
 							{#each sortedFolders as folder}
-								<div class="flex justify-between odd:bg-muted-foreground/5 p-1">
-									<div class="flex items-center gap-0.5 text-sm">
-										<Button
-											size="icon-xs"
-											variant="ghost"
-											class="text-destructive"
-											title="Remove"
-											disabled={deleteNameMutate.isPending}
-											onclick={(e) => {
-												e.stopPropagation();
-												handleDeleteMb(folder.id);
-											}}><Trash /></Button
-										>
-										<p class="font-bold">{folder.name}</p>
-										<p class="text-muted-foreground mb-0.5">({folder.tags.length})</p>
+								{#if folder.userId === userStore.user?.id}
+									<div class="flex justify-between odd:bg-muted-foreground/5 p-1">
+										<div class="flex items-center gap-0.5 text-sm">
+											<Button
+												size="icon-xs"
+												variant="ghost"
+												class="text-destructive"
+												title="Remove"
+												disabled={deleteNameMutate.isPending}
+												onclick={(e) => {
+													e.stopPropagation();
+													handleDeleteMb(folder.id);
+												}}><Trash /></Button
+											>
+											<p class="font-bold">{folder.name}</p>
+											<p class="text-muted-foreground mb-0.5">({folder.tags.length})</p>
+										</div>
+										{#if tagName}
+											<Button size="xs" variant="ghost" onclick={() => handleToggle(folder.id)}>
+												{#if folder.tags.some((t) => t.keyName === tagName)}
+													<span class="font-light">Remove</span>
+													<PinOff class="text-chart-2 size-5" />
+												{:else}
+													<span class="font-light">Add-On</span>
+													<PinIcon class="text-green-400 size-5" />
+												{/if}
+											</Button>
+										{/if}
 									</div>
-									{#if tagName}
-										<Button size="xs" variant="ghost" onclick={() => handleToggle(folder.id)}>
-											{#if folder.tags.some((t) => t.keyName === tagName)}
-												<span class="font-light">Remove</span>
-												<PinOff class="text-chart-2 size-5" />
-											{:else}
-												<span class="font-light">Add-On</span>
-												<PinIcon class="text-green-400 size-5" />
-											{/if}
-										</Button>
-									{/if}
-								</div>
+								{/if}
 							{/each}
 						{/if}
 					</div>
@@ -180,6 +211,47 @@
 							disabled={createNameMutate.isPending}>Create</Button
 						>
 					</form>
+				{:else if tab === 'otherfolders'}
+					{#if allFoldersBm.isSuccess}
+						{#if allFoldersBm.data.filter((f) => f.userId !== userStore.user?.id).length === 0}
+							<NodataTemplate text="No folders..." size="sm" />
+						{/if}
+						{#each sortedFolders as folder}
+							{#if folder.userId !== userStore.user?.id}
+								<div class="flex justify-between odd:bg-muted-foreground/5 p-1">
+									<div class="flex items-center gap-0.5 text-sm">
+										<Button
+											size="icon-xs"
+											variant="ghost"
+											class="text-destructive"
+											title="Remove"
+											disabled={deleteNameMutate.isPending}
+											onclick={(e) => {
+												e.stopPropagation();
+												handleDeleteMb(folder.id);
+											}}><Trash /></Button
+										>
+										<p class="font-bold">{folder.name}</p>
+										<p class="text-muted-foreground mb-0.5">({folder.tags.length})</p>
+										<p class="text-muted-foreground mb-0.5">
+											[{String(folder.user.name).split(' (')[0]}]
+										</p>
+									</div>
+									{#if tagName}
+										<Button size="xs" variant="ghost" onclick={() => handleToggle(folder.id)}>
+											{#if folder.tags.some((t) => t.keyName === tagName)}
+												<span class="font-light">Remove</span>
+												<PinOff class="text-chart-2 size-5" />
+											{:else}
+												<span class="font-light">Add-On</span>
+												<PinIcon class="text-green-400 size-5" />
+											{/if}
+										</Button>
+									{/if}
+								</div>
+							{/if}
+						{/each}
+					{/if}
 				{/if}
 			</section>
 		</div>
